Revamped Conqueror
Now stacks up to 12, granting summon and extra whip damage, plus life regen at max stacks
Also works like the other systems and behaves similarly to new Lethal Tempo

+++ Systems/LethalTempo/LethalTempo.cs
@@ -66,7 +66,7 @@ namespace tsorcRevamp.Systems.LethalTempo
             // Use our animation spritesheet.
             Texture2D ourTexture = animatedTexture.Value;
             // Choose the frame to display, here based on constants and the game's tick count.
-            Rectangle ourSourceRectangle = animatedTexture.Frame(verticalFrames: FrameCount, frameY: (int)modPlayer.Stacks);
+            Rectangle ourSourceRectangle = animatedTexture.Frame(verticalFrames: FrameCount, frameY: modPlayer.Stacks);
 
             // Other stuff you can do in this hook
             /*
